pyproject.toml
[project]
name = "etsp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10.14"
dependencies = [
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "scikit-learn>=1.7.2",
    "soundfile>=0.13.1",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
